Extract common download functions for reuse
